Change repo structure to the more common pkg/cmd structure
