+++ Makefile
@@ -1,7 +1,7 @@
 GIT_REVISION := $(shell git rev-parse HEAD)
 GIT_BRANCH := $(shell git rev-parse --abbrev-ref HEAD)
 
-COLLECTOR_PKG = github.com/ecadlabs/signatory/metrics
+COLLECTOR_PKG = github.com/ecadlabs/signatory/pkg/metrics
 
 signatory:
-	go build -o sig -ldflags "-X $(COLLECTOR_PKG).GitRevision=$(GIT_REVISION) -X $(COLLECTOR_PKG).GitBranch=$(GIT_BRANCH)"
+	go build -ldflags "-X $(COLLECTOR_PKG).GitRevision=$(GIT_REVISION) -X $(COLLECTOR_PKG).GitBranch=$(GIT_BRANCH)" ./cmd/signatory
